🚚 [variables] Move `Binder`, `bind*`, `override` from `templates.domain.binders`
