+++ src/cutty/templates/adapters/questionary/prompts.py
@@ -12,9 +12,9 @@ from pygments.lexers.data import JsonLexer
 from questionary import ValidationError
 from questionary import Validator
 
-from cutty.templates.domain.binders import bind
-from cutty.templates.domain.binders import binddefault
-from cutty.templates.domain.binders import Binder
+from cutty.variables.binders import bind
+from cutty.variables.binders import binddefault
+from cutty.variables.binders import Binder
 from cutty.variables.bindings import Binding
 from cutty.variables.variables import Variable
