+++ tests/unit/templates/adapters/questionary/test_prompts.py
@@ -9,7 +9,7 @@ from prompt_toolkit.input.defaults import create_pipe_input
 from prompt_toolkit.output import DummyOutput
 
 from cutty.templates.adapters.questionary.prompts import createprompt
-from cutty.templates.domain.binders import Binder
+from cutty.variables.binders import Binder
 from cutty.variables.bindings import GenericBinding
 from cutty.variables.variables import GenericVariable
 from tests.util.keys import BACKSPACE
